All fields successfully importing from excel, changed drives int->dec
Dates are now handled correctly, as well as a few other quirks due to Excel storing all numbers as floats. Modified drive fields to be decimal fields instead of integer to properly store the necessary values.

TODO: Properly validate excel sheet to provide useful error messages to the user when attempting to import invalid data. Provide instructions on the import page on the expected format of the excel sheet

+++ infrastructureinventoryproject/infrastructureinventoryapp/models.py
@@ -278,9 +278,9 @@ class ApplicationServer(models.Model):
 
     cpu = models.IntegerField(null=True, blank=True)
     ram = models.IntegerField(null=True, blank=True)
-    c_drive = models.IntegerField(null=True, blank=True)
-    d_drive = models.IntegerField(null=True, blank=True)
-    e_drive = models.IntegerField(null=True, blank=True)
+    c_drive = models.DecimalField(null=True, blank=True, decimal_places=2, max_digits=10)
+    d_drive = models.DecimalField(null=True, blank=True, decimal_places=2, max_digits=10)
+    e_drive = models.DecimalField(null=True, blank=True, decimal_places=2, max_digits=10)
